Beej Server files
The files needed to run the UDP server. It contains the ability to parse CTRL and DATA messages. At the moment, the client sends this message to Beej_Server. "TYPE:CTRL\nPort_Name:A\nPort_Cost:5\nPort_Number:5000\nThis content is data" This can be changed by altering std::String Test_Message on line 79 of Beej_Client or by adding in the commandline entry to sendto on line 96